WIP from functions soeprotocol
